+++ src/app/(content)/phone/page.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 import { ShoppingCart } from 'lucide-react';
 import electronic2 from "../../assets/electronic1.jpg";
 import electronic1 from '../../assets/conditioner1.jpg';
-import StarRating from '../computer/aircondition/starReview';
+import StarRating from '../aircondition/starReview'
 
 // interface Product {
 //   _id: string;
