+++ src/app/(content)/sideBar.tsx
@@ -11,7 +11,7 @@ export const SideBar = () => {
                 <li><Link href="/gadgets">Gadgets</Link></li>
                 <li><Link href="/homeAppliances">Home Appliances</Link></li>
                 <li><Link href="/kitchen">Kitchen Appliances</Link></li>
-                <li><Link href="/computer">PCs & Laptop</Link></li>
+                <li><Link href="/Category/computer">PCs & Laptop</Link></li>
                 <li><Link href="/refregirator">Refrigerator</Link></li>
                 <li><Link href="/phone">Smart Home</Link></li>
             </ul>
