+++ src/app/layout.tsx
@@ -16,7 +16,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <main>{children}</main>
+        {children}
         
         </body>
     </html>
