+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Landing from "@/components/Landing";
 import Sliders from "@/components/Sliders";
-// import Category from "@/components/Category";
+import Category from "@/components/Category";
 import Brand from "@/components/Brand";
 import Footer from "@/components/Footer";
 import Image from "next/image";
@@ -9,7 +9,7 @@ import Image from "next/image";
 export default function Home() {
   return (
     <div className="bg-[E5DFDF]-800">
-   {/* <header>
+   <header>
   <Landing />
    </header>
    
@@ -20,7 +20,6 @@ export default function Home() {
     <Footer />
   </main>
   
-  </main> */}
     </div>
